backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9

# Authentication and Security
python-keycloak==3.3.0
PyJWT[crypto]==2.8.0
cryptography==41.0.4
django-secure==1.0.2
django-csp==3.7
bcrypt==4.0.1
argon2-cffi==23.1.0

# File Handling and Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.3
python-magic==0.4.27
# python-magic-bin==0.4.14; sys_platform == 'win32'
pypdf2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2

# Monitoring and Metrics
django-prometheus==2.3.1
prometheus-client==0.17.1

# Environment Variables
python-dotenv==1.0.0

# WSGI Server
gunicorn==21.2.0

# HTTP and Networking
requests==2.31.0
urllib3
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4

# Django Utils
asgiref==3.7.2
sqlparse==0.4.4
pytz==2023.3.post1
django-filter==23.3
markdown==3.5.1
whitenoise==6.5.0
dj-database-url==2.1.0

# Development and Debug
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.17.2

# AWS SDK Dependencies (for S3)
botocore==1.32.3
jmespath==1.0.1
s3transfer==0.7.0

# Additional Utils
six==1.16.0
typing-extensions==4.8.0

# Cache
django-redis==5.4.0
redis==5.0.1

# Performance and Optimization
django-cleanup==8.0.0
django-imagekit==4.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Documentation
drf-yasg==1.21.7